모니터링 코드 최적화
