+++ src/main/modules/game-monitor/game-monitor.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common'
 import { LoggerModule } from '../logger/logger.module'
+import { MainWindowModule } from '../main-window/main-window.module'
+import { OverlayWindowModule } from '../overlay-window/overlay-window.module'
 import { GameMonitorController } from './game-monitor.controller'
 import { GameMonitorService } from './game-monitor.service'
 
 @Module({
-  imports: [LoggerModule],
+  imports: [LoggerModule, OverlayWindowModule, MainWindowModule],
   controllers: [GameMonitorController],
   providers: [GameMonitorService],
   exports: [GameMonitorService],
