+++ src/main/modules/overlay-window/overlay-window.module.ts
@@ -1,12 +1,9 @@
 import { Module } from '@nestjs/common'
-import { GameMonitorModule } from '../game-monitor/game-monitor.module'
-import { LoggerModule } from '../logger/logger.module'
-import { MainWindowModule } from '../main-window/main-window.module'
 import { OverlayWindowController } from './overlay-window.controller'
 import { OverlayWindowService } from './overlay-window.service'
 
 @Module({
-  imports: [GameMonitorModule, LoggerModule, MainWindowModule],
+  imports: [],
   controllers: [OverlayWindowController],
   providers: [OverlayWindowService],
   exports: [OverlayWindowService],
