+++ src/render/constants/buildInfo.ts
@@ -1,1 +1,1 @@
-export const BUILD_DATE = '202504251321 Client (Vite Developer Build)';
+export const BUILD_DATE = '202504251452 Client (Vite Developer Build)';
